// React/src/model/Item.ts
/**
 * Item
 */
export interface Item {
    /**
    * Item description
    */
    description: string;
    /**
     * Item id
     */
    id?: string;
    /**
     * Item images - relative path
     */
    images?: string[];
    /**
    * Item price
    */
    price: number;
    /**
     * Item title
     */
    title: string;
}